requirements.txt
Flask==1.0.2
python-ldap==3.1.0
